+++ syscall/samples/sched/fork/pidns_init_terminated.c
@@ -21,7 +21,7 @@
 #include <sys/wait.h>
 #include <unistd.h>
 
-#include "helpers.h"
+#include "fork_helpers.h"
 
 
 /**