+++ syscall/samples/sched/namespace/fork_helpers.h
@@ -1,0 +1,1 @@
+../fork/fork_helpers.h