+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
-- syscall: sched: Rename helpers.c to `sched_helpers.c`
+- syscall: fork: Rename `helpers.c` to `fork_helpers.c`
+- syscall: sched: Rename `helpers.c` to `sched_helpers.c`
 - instructions/examples: Add compiler.h
 - module: Add info.sh and kmod/ directory.
 - instructions: Add asm 'ret' test.